Code der GUI aufgeteilt auf einzelne Tabs (Lehrer und Betriebe)
